+++ src/ShapeCrawler/Tables/RightBorder.cs
@@ -60,7 +60,7 @@ internal class RightBorder : IBorder
             this.aTableCellProperties.RightBorderLineProperties.AppendChild(aSolidFill);
         }
         
-        var emus = new Points((decimal)points).AsEmus();
+        var emus = new Points(points).AsEmus();
         this.aTableCellProperties.RightBorderLineProperties!.Width = new Int32Value((int)emus);
     }
